pyproject.toml
[project]
name = "boa-coffee"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.1.0",
    "titanoboa>=0.2.6",
    "vyper>=0.4.1",
]
